pyproject.toml
[tool.poetry]
name = "project-mesdevoirscinq"
version = "0.1.0"
description = "Проект для получения данных о компаниях и вакансиях с hh.ru и загрузки их в PostgreSQL."
authors = ["Annabse27 <annabse27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
tabulate = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
